Write data to a SurveyCTO server

This function updates an existing dataset using a web POST request, as uploading data is not officially supported by the SurveyCTO API.

Usage

scto_write(
  auth,
  data,
  dataset_id,
  dataset_title = dataset_id,
  append = FALSE,
  fill = FALSE
)

Arguments

auth

scto_auth() object.

data

data.frame to upload.

dataset_id

String indicating id of existing dataset.

dataset_title

String indicating title of dataset. Will replace the existing title, regardless of append.

append

Logical indicating whether to append or replace the dataset.

fill

Logical indicating whether to implicitly fill missing columns with NA, i.e., whether to allow a mismatch between columns of the existing dataset and columns of data. Only used if append is TRUE.

Value

A list with elements:

data_old: A data.table of the previous version of the dataset.

response: An object of class httr::response() from the POST request.

Examples

if (FALSE) { # \dontrun{
auth = scto_auth('scto_auth.txt')
r = scto_write(auth, data, 'my_dataset', 'My Dataset')
} # }